* implemented clearCache function in importer to release a specific cached entry from any importer  * [GEAR Studio] implemented effect file selection support

+++ v2.5/GEAR 2.5/RenderComponent.cpp
@@ -116,7 +116,7 @@ void
 G2::RenderComponent::removeAllMeshes()
 {
 	auto* renderSystem = ECSManager::getShared().getSystem<RenderSystem, RenderComponent>();
-	unsigned int numDrawCalls = mDrawCalls.size();
+	unsigned int numDrawCalls = (unsigned int)mDrawCalls.size();
 
 	mDrawCalls.clear();
 	mIaos.clear();
